Code edit tweaks (#75)
* fixed createAssignmentForm closing if wrong input was given

* added timeout to code testing so it does not hang on programs asking for user input and added some visual features to the code editor

* added removal of tmp files for code testing

* Update code_checker.rb

+++ code_checker/code_checker.rb
@@ -6,18 +6,29 @@ class CodeChecker
     OUTPUT = 1
     def self.runCommand(cmd)
         #Runs command and returns if it is successful and stdout and stderr output
-        result = []
-        result, status = Open3.capture2e(cmd)
-        success = (status == 0)
+        puts "Running command: " + cmd
+        result = ""
+        success = -1
+        Open3.popen2e(cmd) { |input, output, wait_thr|
+            begin
+                Timeout.timeout(10) do
+                until output.eof? do
+                    result = output.read
+                end
+            end
+            rescue Timeout::Error
+                # process took longer than 10 seconds
+                Process.kill("KILL", wait_thr.pid)
+                result = "Program took too long to exit."
+            end
+            success = (wait_thr.value == 0)
+        }
         return [success, result]
     end
     
     def self.testCode(language,suid,code)
         ext = FileIO.get_file_extension(language)
         tempFileWithCode = Tempfile.new([suid,ext], :encoding => 'ASCII-8BIT')
-        if (language == "Python")
-            code = "# -*- coding: utf-8 -*-\n" + code
-        end
         code = code.encode('ASCII', invalid: :replace, undef: :replace, replace: "")
         tempFileWithCode.write(code)
         tempFileWithCode.flush()
@@ -26,6 +37,7 @@ class CodeChecker
 
     def self.runProgram(language, filename)
         #Run command to compile and run input program based on language
+        outFileName = ""
         if (language == "Python")
             @compile_result = runCommand("python " + filename)
         elsif (language == "Ruby")
@@ -35,7 +47,6 @@ class CodeChecker
             if (@compile_result[SUCCESS])
                 filename = filename.chomp(".java")
                 @compile_result = runCommand("java " + filename)
-                runCommand("rm " + filename + ".class")
             end
         elsif (language == "C++" || language == "C")
             if (language == "C++")
@@ -47,12 +58,16 @@ class CodeChecker
             end
             if (@compile_result[SUCCESS])
                 @compile_result = runCommand(outFileName)
-                runCommand("rm " + outFileName)
             end      
         else
+            runCommand("rm " + filename);
             return false
         end
         #returns output of program as string
+        runCommand("rm " + filename);
+        if (outFileName != "")
+            runCommand("rm " + outFileName);
+        end
         if (@compile_result[SUCCESS])
             return "Compile Success!\nOutput:\n\n" + @compile_result[OUTPUT]
         else
